refactor(core): extract rotation module from provisioner (S1A-2)
Track A Sprint 1 Step 2. Password rotation moves from provisioner.py to
its own module at src/winpodx/core/rotation/. provisioner.py becomes a
re-export shim so existing imports and test patches keep working until
Step 6 (Sprint 4) deletes the shim.

The module enforces Transport ABC anti-rule #6 in its docstring + via
direct windows_exec.run_in_windows calls — rotation must NEVER route
through Transport since that would tempt callers to use AgentTransport
and expose the new password to the agent process.

Tests for rotation moved to tests/test_rotation/test_rotation.py with
patches retargeted at the new module path. Existing patches via
provisioner.X kept working as a side benefit because the import shim
preserves the public surface.

Tests: 518 passed, 1 skipped (up from 472 before pod/transport/rotation
extractions). ruff clean.
